+++ CHANGELOG.md
@@ -3,9 +3,67 @@ All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
 
+## [0.6.77] - 2025-09-23
+### Changed
+- 补充部分技术文档
+- 修复plantuml语法上的告警
+
+## [0.6.76] - 2025-09-23
+### Changed
+- 新增factors go package概述
+- 调整config c++实现代码的路径
+- update changelog
+
+## [0.6.75] - 2025-09-23
+### Changed
+- 新增检查当前版本是否在pypi中已经存在
+- update changelog
+
+## [0.6.74] - 2025-09-23
+### Changed
+- 优化shell脚本
+- update changelog
+
+## [0.6.73] - 2025-09-23
+### Changed
+- 调整bat发布脚本
+- update changelog
+
+## [0.6.72] - 2025-09-23
+### Changed
+- 修复else不能识别的问题
+- update changelog
+
+## [0.6.71] - 2025-09-23
+### Changed
+- 新增windows ps1的pypi发布脚本
+- update changelog
+
+## [0.6.70] - 2025-09-23
+### Changed
+- quant1x go实现只是提供新扩展的功能，主体功能都在engine中
+- 补全go版本应用入口
+- update changelog
+
+## [0.6.69] - 2025-09-22
+### Changed
+- 调整go应用入口main
+- update changelog
+
+## [0.6.68] - 2025-09-22
+### Changed
+- 计算优化归于标准库组件
+- update changelog
+
+## [0.6.67] - 2025-09-22
+### Changed
+- 更新go依赖库engine版本到1.11.16
+- update changelog
+
 ## [0.6.66] - 2025-09-22
 ### Changed
 - 整理插件模块源文件
+- update changelog
 
 ## [0.6.65] - 2025-09-22
 ### Changed
@@ -529,7 +587,18 @@ All notable changes to this project will be documented in this file.
 - 链接 python win64版本的简易交易客户端
 
 
-[Unreleased]: https://gitee.com/quant1x/quant1x.git/compare/v0.6.66...HEAD
+[Unreleased]: https://gitee.com/quant1x/quant1x.git/compare/v0.6.77...HEAD
+[0.6.77]: https://gitee.com/quant1x/quant1x.git/compare/v0.6.76...v0.6.77
+[0.6.76]: https://gitee.com/quant1x/quant1x.git/compare/v0.6.75...v0.6.76
+[0.6.75]: https://gitee.com/quant1x/quant1x.git/compare/v0.6.74...v0.6.75
+[0.6.74]: https://gitee.com/quant1x/quant1x.git/compare/v0.6.73...v0.6.74
+[0.6.73]: https://gitee.com/quant1x/quant1x.git/compare/v0.6.72...v0.6.73
+[0.6.72]: https://gitee.com/quant1x/quant1x.git/compare/v0.6.71...v0.6.72
+[0.6.71]: https://gitee.com/quant1x/quant1x.git/compare/v0.6.70...v0.6.71
+[0.6.70]: https://gitee.com/quant1x/quant1x.git/compare/v0.6.69...v0.6.70
+[0.6.69]: https://gitee.com/quant1x/quant1x.git/compare/v0.6.68...v0.6.69
+[0.6.68]: https://gitee.com/quant1x/quant1x.git/compare/v0.6.67...v0.6.68
+[0.6.67]: https://gitee.com/quant1x/quant1x.git/compare/v0.6.66...v0.6.67
 [0.6.66]: https://gitee.com/quant1x/quant1x.git/compare/v0.6.65...v0.6.66
 [0.6.65]: https://gitee.com/quant1x/quant1x.git/compare/v0.6.64...v0.6.65
 [0.6.64]: https://gitee.com/quant1x/quant1x.git/compare/v0.6.63...v0.6.64